Fix Comment-Class ID mapping and refactor data management
- Fix Class upload to use original IDs from JSON for proper Comment mapping
- Comments now properly display for each Class (20 comments for Class '1')
- Refactor data management functions to separate utility file
- Improve delete logic with batch processing and accurate progress tracking
- Remove Korean comments and clean up code structure
- Add detailed logging for better debugging
